app/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
diff-match-patch==20230430
Django==4.2.3
django-import-export==4.1.0
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.7
openpyxl==3.1.5
packaging==24.1
pillow==10.2.0
psycopg2-binary==2.9.6
pypng==0.20220715.0
qrcode==7.4.2
requests==2.31.0
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.2
urllib3==2.2.2
xlrd==2.0.1
xlwt==1.3.0
django-admin-interface==0.26.0
dj-user-login-history==1.0.5
django-request==1.6.3
django-admin-charts==1.3.1
django-light==0.1.0.post3
django-defender==0.9.8
django-admin-logs==1.2.0
django-nested-inline==0.4.6
django-admin-sortable2==2.2.3
django-image-uploader-widget==1.0.0